requirements.txt
Flask==0.10.1
Flask-SocketIO==2.1
eventlet-0.18.3
greenlet-0.4.9
